refactor(webserver): extract juniper-axum handler to webserver/axum
