+++ crates/juniper-axum/Cargo.toml
@@ -8,6 +8,5 @@ homepage.workspace = true
 [dependencies]
 axum.workspace = true
 juniper.workspace = true
-juniper_graphql_ws = "0.3.0"
 serde.workspace = true
 serde_json.workspace = true
