+++ ee/tabby-webserver/Cargo.toml
@@ -23,6 +23,7 @@ hyper = { workspace = true, features = ["client"] }
 jsonwebtoken = "9.1.0"
 juniper.workspace = true
 juniper-axum = { path = "../../crates/juniper-axum" }
+juniper_graphql_ws = "0.3"
 lazy_static.workspace = true
 lettre = { version = "0.11.3", features = ["tokio1", "tokio1-native-tls"] }
 mime_guess = "2.0.4"
