+++ ee/tabby-webserver/src/hub/mod.rs
@@ -13,7 +13,6 @@ use axum::{
     TypedHeader,
 };
 use hyper::{Body, StatusCode};
-use juniper_axum::extract::AuthBearer;
 use tabby_common::{
     api::{code::SearchResponse, event::LogEntry},
     config::RepositoryConfig,
@@ -22,7 +21,7 @@ use tarpc::server::{BaseChannel, Channel};
 use tracing::warn;
 use websocket::WebSocketTransport;
 
-use crate::schema::ServiceLocator;
+use crate::{axum::extract::AuthBearer, schema::ServiceLocator};
 
 pub(crate) async fn ws_handler(
     ws: WebSocketUpgrade,
