+++ ee/tabby-webserver/src/lib.rs
@@ -1,5 +1,6 @@
 //! Defines behavior for the tabby webserver which allows users to interact with enterprise features.
 //! Using the web interface (e.g chat playground) requires using this module with the `--webserver` flag on the command line.
+mod axum;
 mod cron;
 mod handler;
 mod hub;
