+++ ee/tabby-webserver/src/schema/mod.rs
@@ -21,10 +21,7 @@ use juniper::{
     graphql_object, graphql_value, EmptySubscription, FieldError, FieldResult, GraphQLObject,
     IntoFieldError, Object, RootNode, ScalarValue, Value, ID,
 };
-use juniper_axum::{
-    relay::{self, Connection},
-    FromAuth,
-};
+use juniper_axum::relay::{self, Connection};
 use tabby_common::api::{code::CodeSearch, event::EventLogger};
 use tracing::error;
 use validator::{Validate, ValidationErrors};
@@ -45,7 +42,7 @@ use self::{
         NetworkSetting, NetworkSettingInput, SecuritySetting, SecuritySettingInput, SettingService,
     },
 };
-use crate::schema::repository::FileEntrySearchResult;
+use crate::{axum::FromAuth, schema::repository::FileEntrySearchResult};
 
 pub trait ServiceLocator: Send + Sync {
     fn auth(&self) -> Arc<dyn AuthenticationService>;
